use snapdragon, clean up
