+++ .travis.yml
@@ -1,6 +1,7 @@
 sudo: false
 language: node_js
 node_js:
+  - 'node'
   - '6'
   - '4'
   - '0.12'
