--- lib/extglob/compiler.js
@@ -1,134 +1,0 @@
-'use strict';
-
-var debug = require('debug')('extglob:compiler');
-var utils = require('../utils');
-
-/**
- * Create a new `Compiler` with the given `options`.
- * @param {Object} `options`
- */
-
-function Compiler(options) {
-  debug('compiler', __filename);
-  this.options = options || {};
-  this.compilers = {};
-  this.output = '';
-}
-
-/**
- * Prototype methods
- */
-
-Compiler.prototype = {
-
-  /**
-   * Throw an error message with details including the cursor position.
-   * @param {String} `msg` Message to use in the Error.
-   */
-
-  error: function(msg, node) {
-    var pos = node.position;
-    var message = this.options.source + ' column:' + pos.start.column + ': ' + msg;
-
-    var err = new Error(message);
-    err.reason = msg;
-    err.column = pos.start.column;
-    err.source = this.pattern;
-
-    if (this.options.silent) {
-      this.errors.push(err);
-    } else {
-      throw err;
-    }
-  },
-
-  /**
-   * Emit `node.val`
-   */
-
-  emit: function(str) {
-    this.output += str;
-    return str;
-  },
-
-  /**
-   * Add a compiler `fn` with the given `name`
-   */
-
-  set: function(name, fn) {
-    this.compilers[name] = fn;
-    return this;
-  },
-
-  /**
-   * Get the previous AST node.
-   */
-
-  prev: function(n) {
-    return this.ast.nodes[this.idx - (n || 1)] || { type: 'bos' };
-  },
-
-  /**
-   * Get the next AST node.
-   */
-
-  next: function(n) {
-    return this.ast.nodes[this.idx + (n || 1)] || { type: 'eos' };
-  },
-
-  /**
-   * Visit `node`.
-   */
-
-  visit: function(node, nodes, i) {
-    var fn = this.compilers[node.type];
-    this.idx = i;
-
-    if (typeof fn !== 'function') {
-      throw this.error('compiler "' + node.type + '" is not registered', node);
-    }
-    return fn.call(this, node, nodes, i);
-  },
-
-  /**
-   * Map visit over array of `nodes`.
-   */
-
-  mapVisit: function(nodes) {
-    var len = nodes.length;
-    var idx = -1;
-    while (++idx < len) {
-      this.visit(nodes[idx], nodes, idx);
-    }
-    return this;
-  },
-
-  /**
-   * Compile `ast`.
-   */
-
-  compile: function(ast, options) {
-    var opts = utils.extend({}, this.options, options);
-    this.ast = ast;
-    this.parsingErrors = this.ast.errors;
-
-    // source map support
-    if (opts.sourcemap) {
-      var sourcemaps = require('./source-maps');
-      sourcemaps(this);
-      this.mapVisit(this.ast.nodes);
-      this.applySourceMaps();
-      this.map = opts.sourcemap === 'generator' ? this.map : this.map.toJSON();
-      return this;
-    }
-
-    this.mapVisit(this.ast.nodes);
-    return this;
-  }
-};
-
-/**
- * Expose `Compiler`
- */
-
-module.exports = Compiler;
