+++ lib/utils.js
@@ -2,19 +2,12 @@
 
 exports.define = require('define-property');
 exports.extend = require('extend-shallow');
-exports.SourceMap = require('source-map');
-exports.sourceMapResolve = require('source-map-resolve');
-exports.urix = require('urix');
 
 /**
- * Get the last element from `array`
- * @param {Array} `array`
- * @return {*}
+ * Create a negation regex from the given string
+ * @param {String} `str`
+ * @return {RegExp}
  */
-
-exports.last = function(arr) {
-  return arr[arr.length - 1];
-};
 
 exports.not = function(str) {
   return '^((?!(?:' + str + ')).)*';
